You are provided with two sorted integer arrays, nums1 and nums2, along with two integers, m and n, which represent the count of elements in nums1 and nums2, respectively.
Your task is to efficiently merge the elements from nums2 into nums1, resulting in a single array that is sorted in non-decreasing order. The merged array should be stored in nums1. nums1 has been allocated enough space to accommodate both arrays, where the initial n elements at the end of nums1 are set to 0 and should be disregarded. nums2 has a length of n.